requirements.txt
fastapi
uvicorn
python-multipart
easyocr
transformers
torch
numpy
Pillow
opencv-python
google-generativeai
python-dotenv
ollama
